refactor(tools/distro-info): for later use in ubuntu and debian versioning
cr changes and migrated from mjs to ts
