+++ lib/modules/versioning/ubuntu/index.ts
@@ -1,5 +1,5 @@
 import { regEx } from '../../../util/regex';
-import DistroInfo from '../distro';
+import { DistroInfo } from '../distro';
 import type { NewValueConfig, VersioningApi } from '../types';
 
 export const id = 'ubuntu';
